- Skip integration tests for now, can be done on Travis if credentials are sent securely - Refactor a bit so that HealthBee is one entity - Expand documentation
