+++ README.md
@@ -1,36 +1,35 @@
 [![Build Status](https://travis-ci.com/dnataraj/healthbee.svg?token=g7PAjZdpVPTj6UWnsEsA&branch=main)](https://travis-ci.com/dnataraj/healthbee)
 
+HealthBee is a website availability monitoring service.
+
+####System Requirements
+* HealthBee has been tested to work on Ubuntu 20.04 LTS with Golang 15.2
+
+####User Guide
+
+The HealthBee API can be used to register any (web) site for monitoring. The API is further specified in detail below.
+
+* ```GET /sites``` : Lists all the currently registered sites
+* ```POST /sites/{id}/stop``` : Stops monitoring activity for a particular site
+* ```POST /sites``` : Register a new site for monitoring
+    * The request for site registration can be specified in JSON as follows :
+        ```
+            {   
+                "url": "https://www.google.com", <-- the site address 
+                "interval": "4s",  <-- a monitoring interval, in seconds
+                "pattern": "content"  <-- an optional regular expression that is searched for in the returned page
+            }
+        ```
+* ```GET /sites/{id}/metrics``` will return the last 20 metrics for the given site in JSON 
+
+##### Installation and setup
+
+* HealthBee can be installed on your system using the ```go install``` command, for example
+    ```> go install github.com/dnataraj/healthbee```
+* This will install HealthBee binary to a directory specified in the ```GOBIN``` env var
+* A clean shut down of the service can be achieved by simply doing a Ctrl-C
+
 #### Development Notes
 
-* TODO: Complete testing - unit and integration
-* TODO: Possibly combine monitor and auditor for simpler configuration and running
-* TODO: Add stop and deregistration
-
-#### Usage
-
-* Start HealthBee API service
-    - First start the monitoring API
-      - ```go run ./cmd/web --brokers "<kafka service endpoint>" --dsn "<postgres connection string>"```
-      - ```go run ./cmd/web -h``` provides usage information
-      - Local brokers and databases can be used, just provide the local service addressed and use the ```-local``` flag
-    - Then, start the auditor service
-      - ```go run ./cmd/auditors --brokers "<kafka service endpoint>" --dsn "<postgres connection string>"```
-      - Currently there are 2 consumers for the auditor service
-    
-* Register a site for monitoring
-    - This can be done with curl and triggers site monitoring at the configured interval e.g.
-
-```
-   curl --request POST \
-  --url http://localhost:8000/monitor \
-  --header 'content-type: application/json' \
-  --data '{
-	"url": "https://www.duckduckgo.com",
-	"interval": "5s",
-	"pattern": "privacy"
-  }'
-
-```
-    
-    
-* A clean shut down of both the monitor and the auditor can be achieved by simply doing a Ctrl-C
+* TODO: Highlight testing strategy and possibilities - both unit and integration
+* TODO: Support API
