+++ cmd/healthbee/main.go
@@ -9,6 +9,7 @@ import (
 	"github.com/dnataraj/healthbee/pkg/models/postgres"
 	_ "github.com/lib/pq"
 	"github.com/segmentio/kafka-go"
+	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
@@ -22,7 +23,9 @@ type application struct {
 	errorLog *log.Logger
 	infoLog  *log.Logger
 
-	sites    *postgres.SiteModel
+	sites   *postgres.SiteModel
+	results *postgres.ResultModel
+
 	monitors map[int]*pkg.Monitor
 	writer   *kafka.Writer
 	wg       *sync.WaitGroup
@@ -31,6 +34,7 @@ type application struct {
 
 func main() {
 	local := flag.Bool("local", false, "Set for local development mode")
+	reset := flag.Bool("reset", false, "Clear all site registrations and metrics")
 	addr := flag.String("addr", ":8000", "HTTP network address")
 	brokerList := flag.String("brokers", "localhost:9092", "Comma separated distributed cache peers")
 	dsn := flag.String("dsn", "", "DSN/Connection string for the PostgreSQL database")
@@ -47,6 +51,19 @@ func main() {
 		errorLog.Fatal("server: unable to connect to sites database: ", err.Error())
 	}
 	defer db.Close()
+
+	if *reset {
+		// re-initialize the database
+		script, err := ioutil.ReadFile("pkg/models/postgres/testdata/setup.sql")
+		if err != nil {
+			errorLog.Fatal("server: unable to reset HealthBee database: ", err.Error())
+		}
+		infoLog.Println("server: HealthBee database reset requested, performing...")
+		_, err = db.Exec(string(script))
+		if err != nil {
+			errorLog.Fatal("server: unable to reset HealthBee database: ", err.Error())
+		}
+	}
 
 	// initialize TLS config for non-local services
 	var tlsConfig *tls.Config
@@ -75,6 +92,7 @@ func main() {
 		errorLog: errorLog,
 		infoLog:  infoLog,
 		sites:    &postgres.SiteModel{DB: db},
+		results:  &postgres.ResultModel{DB: db},
 		monitors: make(map[int]*pkg.Monitor),
 		writer:   w,
 		wg:       &wg,
@@ -88,6 +106,17 @@ func main() {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
+	// start the auditors
+	infoLog.Println("server: creating readers for incoming metrics...")
+	dialer := &kafka.Dialer{Timeout: 10 * time.Second, TLS: tlsConfig}
+	r1 := newReader(brokers, dialer)
+	r2 := newReader(brokers, dialer)
+	infoLog.Println("auditor: starting 2 readers for incoming metrics...")
+	wg.Add(1)
+	go app.read(ctx, 1, r1, &wg)
+	wg.Add(1)
+	go app.read(ctx, 2, r2, &wg)
+
 	infoLog.Printf("starting HealthBee API server on %s", *addr)
 	wg.Add(1)
 	go webServer(ctx, srv, &wg)
@@ -96,7 +125,7 @@ func main() {
 	c := make(chan os.Signal, 1)
 	signal.Notify(c, os.Interrupt)
 	<-c
-	infoLog.Println("server: shutting down HTTP services...")
+	infoLog.Println("server: shutting down HTTP and auditor services...")
 	cancel()
 
 	// cancel all monitors
@@ -124,4 +153,13 @@ func webServer(ctx context.Context, srv *http.Server, wg *sync.WaitGroup) {
 	shCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
 	_ = srv.Shutdown(shCtx)
+}
+
+func newReader(brokers []string, dialer *kafka.Dialer) *kafka.Reader {
+	return kafka.NewReader(kafka.ReaderConfig{
+		Brokers: brokers,
+		GroupID: "message-reader-group",
+		Topic:   "Metrics",
+		Dialer:  dialer,
+	})
 }
